Refactored UI. Renamed UI elements and moved activities into their own package.
