+++ app/src/main/res/layout/activity_humidity_display.xml
@@ -4,43 +4,53 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.teamplantpower.team_plant_power.HumidityDisplay">
+    tools:context="com.teamplantpower.activities.HumidityDisplay">
 
     <TextView
         android:id="@+id/humidityExposureValue"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:text="@string/value_humidity"
-        app:layout_constraintBottom_toTopOf="@+id/message"
-        android:layout_marginStart="51dp"
+        android:layout_marginStart="100dp"
         app:layout_constraintTop_toTopOf="parent"
-        android:layout_marginTop="55dp"
-        android:layout_marginLeft="51dp"
-        app:layout_constraintVertical_bias="0.277" />
+        android:layout_marginTop="65dp"
+        android:layout_marginLeft="76dp"
+        app:layout_constraintLeft_toLeftOf="parent" />
 
     <TextView
-        android:id="@+id/percent"
+        android:id="@+id/percentMax"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
+        android:layout_marginTop="282dp"
         android:text="%"
-        app:layout_constraintLeft_toLeftOf="parent"
-        app:layout_constraintRight_toRightOf="parent"
         app:layout_constraintHorizontal_bias="0.919"
-        app:layout_constraintTop_toTopOf="parent"
-        android:layout_marginTop="282dp" />
-
-    <TextView
-        android:id="@+id/textView5"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:text="%"
         app:layout_constraintLeft_toLeftOf="parent"
         app:layout_constraintRight_toRightOf="parent"
-        android:layout_marginTop="55dp"
         app:layout_constraintTop_toTopOf="parent" />
 
+    <TextView
+        android:id="@+id/percentMin"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_marginTop="282dp"
+        android:text="%"
+        app:layout_constraintHorizontal_bias="0.457"
+        app:layout_constraintLeft_toLeftOf="parent"
+        app:layout_constraintRight_toRightOf="parent"
+        app:layout_constraintTop_toTopOf="parent" />
+
+    <TextView
+        android:id="@+id/percentValue"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:text="%"
+        app:layout_constraintTop_toTopOf="parent"
+        android:layout_marginTop="65dp"
+        android:layout_marginRight="187dp"
+        app:layout_constraintRight_toRightOf="parent" />
+
     <Button
-        android:id="@+id/refreshHumidityExposure"
+        android:id="@+id/refreshHumidity"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:layout_marginTop="123dp"
@@ -76,9 +86,8 @@
         android:layout_marginStart="30dp" />
 
 
-
     <Button
-        android:id="@+id/setMin_MaxHumidity"
+        android:id="@+id/setHumidity"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:layout_marginTop="373dp"
