+++ app/src/main/res/layout/activity_light_display.xml
@@ -4,54 +4,53 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.teamplantpower.team_plant_power.LightDisplay">
+    tools:context="com.teamplantpower.activities.LightDisplay">
 
     <TextView
-        android:id="@+id/textView4"
+        android:id="@+id/percentMin"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
+        android:layout_marginTop="282dp"
         android:text="%"
+        app:layout_constraintHorizontal_bias="0.473"
         app:layout_constraintLeft_toLeftOf="parent"
         app:layout_constraintRight_toRightOf="parent"
-        app:layout_constraintHorizontal_bias="0.473"
-        app:layout_constraintTop_toTopOf="parent"
-        android:layout_marginTop="282dp" />
+        app:layout_constraintTop_toTopOf="parent" />
 
     <TextView
         android:id="@+id/lightExposureValue"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:layout_marginLeft="58dp"
         android:text="@string/value_light"
-        app:layout_constraintLeft_toLeftOf="parent"
-        app:layout_constraintBottom_toTopOf="@+id/message"
-        android:layout_marginStart="51dp"
+        android:layout_marginStart="60dp"
         app:layout_constraintTop_toTopOf="parent"
-        app:layout_constraintVertical_bias="0.277" />
+        android:layout_marginTop="62dp"
+        android:layout_marginLeft="98dp"
+        app:layout_constraintLeft_toLeftOf="parent" />
 
     <TextView
-        android:id="@+id/percent"
+        android:id="@+id/percentMax"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
+        android:layout_marginTop="282dp"
         android:text="%"
-        app:layout_constraintLeft_toLeftOf="parent"
-        app:layout_constraintRight_toRightOf="parent"
         app:layout_constraintHorizontal_bias="0.919"
-        app:layout_constraintTop_toTopOf="parent"
-        android:layout_marginTop="282dp" />
-
-    <TextView
-        android:id="@+id/textView5"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:text="%"
         app:layout_constraintLeft_toLeftOf="parent"
         app:layout_constraintRight_toRightOf="parent"
-        android:layout_marginTop="55dp"
+        app:layout_constraintTop_toTopOf="parent" />
+
+    <TextView
+        android:id="@+id/percentValue"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_marginRight="187dp"
+        android:layout_marginTop="62dp"
+        android:text="%"
+        app:layout_constraintRight_toRightOf="parent"
         app:layout_constraintTop_toTopOf="parent" />
 
     <Button
-        android:id="@+id/refreshLightExposure"
+        android:id="@+id/refreshLight"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:layout_marginTop="123dp"
@@ -87,7 +86,7 @@
         android:layout_marginStart="30dp" />
 
     <Button
-        android:id="@+id/setMin_MaxLight"
+        android:id="@+id/setLight"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:layout_marginTop="373dp"
