+++ app/src/main/res/layout/activity_temperature_display.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.teamplantpower.team_plant_power.TemperatureDisplay">
+    tools:context="com.teamplantpower.activities.TemperatureDisplay">
 
     <TextView
         android:id="@+id/temperatureTitle"
@@ -17,47 +17,47 @@
         app:layout_constraintLeft_toLeftOf="parent" />
 
     <TextView
-        android:id="@+id/celsius"
+        android:id="@+id/minCelsius"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:text="'C"
-        app:layout_constraintTop_toTopOf="parent"
-        android:layout_marginTop="356dp"
         android:layout_marginLeft="177dp"
-        app:layout_constraintLeft_toLeftOf="parent" />
-
-    <TextView
-        android:id="@+id/textView10"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:text="'C"
-        android:layout_marginRight="16dp"
-        app:layout_constraintRight_toRightOf="parent"
-        app:layout_constraintTop_toTopOf="parent"
         android:layout_marginTop="356dp"
-        app:layout_constraintLeft_toRightOf="@+id/celsius"
+        android:text="@string/text_degree_c"
+        app:layout_constraintLeft_toLeftOf="parent"
+        app:layout_constraintTop_toTopOf="parent" />
+
+    <TextView
+        android:id="@+id/degreeCMax"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
         android:layout_marginLeft="8dp"
-        app:layout_constraintHorizontal_bias="1.0" />
+        android:layout_marginRight="16dp"
+        android:layout_marginTop="356dp"
+        android:text="@string/text_degree_c"
+        app:layout_constraintHorizontal_bias="1.0"
+        app:layout_constraintLeft_toRightOf="@+id/minCelsius"
+        app:layout_constraintRight_toRightOf="parent"
+        app:layout_constraintTop_toTopOf="parent" />
 
     <TextView
-        android:id="@+id/textView11"
+        android:id="@+id/degreeCValue"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:text="'C"
-        app:layout_constraintTop_toTopOf="parent"
+        android:layout_marginRight="93dp"
         android:layout_marginTop="105dp"
-        app:layout_constraintLeft_toRightOf="@+id/celciusValue"
-        android:layout_marginLeft="187dp" />
+        android:text="@string/text_degree_c"
+        app:layout_constraintRight_toRightOf="parent"
+        app:layout_constraintTop_toTopOf="parent" />
 
     <TextView
-        android:id="@+id/fahrenheit"
+        android:id="@+id/degreeFValue"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:text="'F"
-        app:layout_constraintLeft_toRightOf="@+id/textView2"
-        android:layout_marginLeft="187dp"
-        app:layout_constraintTop_toTopOf="parent"
-        android:layout_marginTop="151dp" />
+        android:layout_marginRight="95dp"
+        android:layout_marginTop="144dp"
+        android:text="@string/text_degree_f"
+        app:layout_constraintRight_toRightOf="parent"
+        app:layout_constraintTop_toTopOf="parent" />
 
     <TextView
         android:id="@+id/celciusValue"
@@ -70,17 +70,17 @@
         app:layout_constraintLeft_toLeftOf="parent" />
 
     <TextView
-        android:id="@+id/textView2"
+        android:id="@+id/fahrenheitValue"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:text="@string/value_fahrenheit"
-        app:layout_constraintTop_toTopOf="parent"
-        android:layout_marginTop="151dp"
         android:layout_marginLeft="65dp"
-        app:layout_constraintLeft_toLeftOf="parent" />
+        android:layout_marginTop="151dp"
+        android:text="@string/value_fahrenheit"
+        app:layout_constraintLeft_toLeftOf="parent"
+        app:layout_constraintTop_toTopOf="parent" />
 
     <Button
-        android:id="@+id/button"
+        android:id="@+id/refreshTemperature"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:layout_marginTop="217dp"
@@ -115,7 +115,7 @@
         app:layout_constraintLeft_toLeftOf="parent" />
 
     <Button
-        android:id="@+id/button6"
+        android:id="@+id/setTemperature"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:layout_marginTop="429dp"
@@ -126,13 +126,13 @@
         app:layout_constraintTop_toTopOf="parent" />
 
     <TextView
-        android:id="@+id/textView12"
+        android:id="@+id/message"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
+        android:layout_marginLeft="33dp"
         android:layout_marginTop="293dp"
         android:text="Set min and max temperature:"
-        app:layout_constraintTop_toTopOf="parent"
-        android:layout_marginLeft="33dp"
-        app:layout_constraintLeft_toLeftOf="parent" />
+        app:layout_constraintLeft_toLeftOf="parent"
+        app:layout_constraintTop_toTopOf="parent" />
 
 </android.support.constraint.ConstraintLayout>
